requirements.txt
biopython
numba